Demonstrate version bug
When tags are not ordered from highest to lowest, the returned latest
version is incorrect. The version will always be the first tag if it
passes the version constraint, which by default is simply `>= 0.0.0`.
Even if it is manually specified in the upstream config to match the
current version, the returned latest version can still be incorrect when
there are multiple versions, ordered "randomly", higher than the current
one.

A fix would need to iterate all tags, compare any tag with the
previously highest tag, and then return the last one.